requirements.txt
numpy==1.22.0
scikit-learn==1.0.2
streamlit==1.5.1
tensorflow==2.7.0
Pillow==9.0.0
